DiceChecker version 6.0.0.1 - release
DiceChecker version 6.0.0.1 - release (added physicalCryptoRandomStream to deal with physical RAM)

+++ DiceChecker/blockFrequencyTest.h
@@ -1,8 +1,8 @@
 //
 // Creator:    http://www.dicelocksecurity.com
-// Version:    vers.5.0.0.1
+// Version:    vers.6.0.0.1
 //
-// Copyright © 2008-2011 DiceLock Security, LLC. All rights reserved.
+// Copyright © 2008-2012 DiceLock Security, LLC. All rights reserved.
 //
 //                               DISCLAIMER
 //
@@ -79,7 +79,7 @@ namespace DiceLockSecurity {
 			CLASS_DECLSPEC unsigned long int GetMinimumLength(void);
 
 			/// Sets the blockLength
-			CLASS_DECLSPEC void SetBlockLength(unsigned int);
+			CLASS_DECLSPEC void SetBlockLength(unsigned long int);
 
 			/// Gets the blockLength
 			CLASS_DECLSPEC unsigned long int GetBlockLength(void);
